+++ src/actions/index.js
@@ -1,12 +1,12 @@
 import {FETCH_PHOTOS_TO_MODERATE} from "./actionTypes";
-import config from "../custom/config";
+import dbFirebase from "../dbFirebase";
 
 let unregisterPhotosToModerateObserver = undefined;
 
 export const startFetchingPhotosToModerate = (dispatch) => async () => {
   if (!unregisterPhotosToModerateObserver) {
 
-    unregisterPhotosToModerateObserver = config.dbModule.onPhotosToModerate(photosToModerate => {
+    unregisterPhotosToModerateObserver = dbFirebase.onPhotosToModerate(photosToModerate => {
       dispatch({
         type: FETCH_PHOTOS_TO_MODERATE,
         payload: photosToModerate