+++ src/custom/config.js
@@ -1,24 +1,8 @@
-import authFirebase from './authFirebase'
-import LoginFirebase from "./components/LoginFirebase";
-import dbFirebase from "./dbFirebase";
-import AnonymousPage from "./components/AnonymousPage";
-import SignedinPage from "./components/SignedinPage";
-import ModeratorPage from "../components/ModeratorPage";
-import EverybodyPage from "./components/EverybodyPage";
-
 import styles from './config.scss';
 const primaryColor = styles.primary;
 const secondaryColor = styles.secondary;
 
 export default {
-  uploadPhoto: dbFirebase.uploadPhoto,
-  loginComponent: LoginFirebase,
-  authModule: authFirebase,
-  dbModule: dbFirebase,
-  AnonymousPage: AnonymousPage,
-  SignedinPage: SignedinPage,
-  ModeratorPage: ModeratorPage,
-  EverybodyPage: EverybodyPage,
   MAX_IMAGE_SIZE: 2048,
   THEME: {
     typography: {