+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - recognize moderator
 - introducing configurable components
 - get user groups
+- moderator user
 
 ## 0.2.2 (2018-11-05)
 - fix map iOS