+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - login
 - recognize moderator
 - introducing configurable components
+- get user groups
+- moderator user
 
 ## 0.2.2 (2018-11-05)
 - fix map iOS